Move ssdp functions to shared
